GSYE-295: Move market related utils in separate module in order to reduce size of gsy_e_core/util.py
